pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling" ]

[project]
name = "exif-stripper"
version = "1.0.0"
description = "An easy-to-use tool to ensure image EXIF metadata is removed."
readme = "README.md"
keywords = [
  "exif",
  "image processing",
  "metadata",
  "pre-commit",
]
license = { file = "LICENSE" }
authors = [
  { name = "Stefanie Molin", email = "exif-stripper@stefaniemolin.com" },
]
requires-python = ">=3.11"
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

dependencies = [
  "pillow>=11.0.0", # Python 3.9 - 3.13 (https://pillow.readthedocs.io/en/stable/installation/python-support.html)
]

urls.Documentation = "https://github.com/stefmolin/exif-stripper"
urls.Homepage = "https://github.com/stefmolin/exif-stripper"
urls.Source = "https://github.com/stefmolin/exif-stripper"
scripts.exif-stripper = "exif_stripper.cli:main"

[dependency-groups]
dev = [ "pre-commit", { include-group = "test" } ]

test = [
  "pytest>=8.3.4",
  "pytest-cov>=4.1.0",
  "pytest-randomly>=3.12.0",
]

[tool.ruff]
line-length = 88
format.indent-style = "space"
format.quote-style = "single"
format.docstring-code-format = true
lint.select = [
  "ANN",  # flake8-annotations
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "E",    # pycodestyle error
  "ERA",  # eradicate (commented out code)
  "F",    # pyflakes
  "FA",   # flake8-future-annotations
  "I",    # isort
  "ICN",  # flake8-import-conventions
  "N",    # pep8-naming
  "PERF", # perflint
  "PT",   # flake8-pytest-style
  "PTH",  # flake8-use-pathlib
  "RET",  # flake8-return
  "RSE",  # flake8-raise
  "RUF",  # ruff-specific rules
  "SIM",  # flake8-simplify
  "TC",   # flake8-type-checking (performance improvements)
  "TRY",  # tryceratops
  "UP",   # pyupgrade
  "W",    # pycodestyle warning
]
lint.ignore = [
  "E501", # line-too-long
]

lint.extend-per-file-ignores."tests/*" = [ "ANN" ] # don't require annotations for tests
lint.isort.known-first-party = [
  "exif_stripper",
]

[tool.pytest.ini_options]
addopts = [
  "-ra",
  "-l",
  "-v",
  "--tb=short",
  "--import-mode=importlib",
  "--strict-markers",
  "--strict-config",
  "--cov=exif_stripper",
  "--cov=tests",
  "--no-cov-on-fail",
  "--cov-report=term-missing",
]
xfail_strict = true
testpaths = [
  "tests",
]

[tool.numpydoc_validation]
checks = [
  "all",  # report on all checks
  "ES01", # but don't require an extended summary
  "EX01", # or examples
  "SA01", # or a see also section
  "SS06", # and don't require the summary to fit on one line
]
exclude = [ # don't report on checks for these
  '\.__init__$',
  '\.__repr__$',
  '\.__str__$',
]
override_SS05 = [ # allow docstrings to start with these words
  '^Process ',
  '^Assess ',
  '^Access ',
]
